twitter_api_server/requirements.txt
anyio==3.6.1
fastapi==0.79.0
idna==3.3
pydantic==1.9.2
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn[standard]
python-dotenv
tweepy
isort